requirements.txt
gitpython==3.1.42
python-dateutil==2.8.2 